--- a/test/Project.toml
+++ b/test/Project.toml
@@ -1,9 +1,7 @@
 [deps]
 Compose = "a81c6b42-2e10-5240-aca2-a61377ecd94b"
-POMDPs = "a93abf59-7444-517b-a68a-c42f96afdd7d"
 POMDPTools = "7588e00f-9cae-40de-98dc-e0c70c48cdd7"
+POMDPs = "a93abf59-7444-517b-a68a-c42f96afdd7d"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[compat]
-POMDPTools = "0.1"