@zsunberg 2023-05-23
This also works nowadays: #94 :)