@WhiffleFish 2023-05-23
Woah is that documented anywhere?

Why does that work? Compiler magic?