@zsunberg 2023-05-23
I'm not sure where/if it is documented, but I think the compiler is now able to avoid allocations for any mutable object that does not ever leave the function it is allocated in. I found it out from @narijauskas last week.

Should I just merge #94 in lieu of this, or is there any reason to still consider this one?