@WhiffleFish 2023-05-24
Hmm okay, it doesn't seem to be the case for *any* mutable object. But yes, absolutely just merge #94. It achieves everything this PR aims to do, but in a significantly simpler way.